requirements.txt
python-telegram-bot
requests
spotipy
dotenv
